Rework expression and variable reference grammars.
The expression grammar is now expressed as a single left recursive rule,
relying on antlr4 to get the precedence right. Also made it more similar
to the Verilog standard expression grammar. This required factoring out
variable references as a separate rule (var_ref), this will also help
with typing in the future and makes the grammar more explicit.

Due to this rework, there is now one piece of former syntax which is no
longer syntactically correct. These are applying '[ ]' to an arbitrary
subexpression, in particular a .read() call. For example, this will no
longer parse:

in sync u10 p_in;

u9 val;

val = p_in.read()[9:1];

These are trivial to rewrite as:

u1 unused;

{ val, unused } = p_in.read();

There are 2 instances of this in the test set, so it should not be a big
issue. The benefit however that it is guaranteed that you will not be
able to write an expression that is not correct in Verilog.

Of course you can still write p_in[9:1], to access a port signal.

For this work, it was also required to adjust assignment statements
(including everything that desugars into an assignment statement
like ++ -- +=), which now only accept lvalues that the Verilog standard
accepts (that is variable references, or a concatenation of lvalues).

Streamlining the grammar also has the unexpected side effect of making
the whole compiler 30% faster!