Make parameter initialiser mandatory.
Optional parameter initializers are only supported in SystemVerilog-2009
or later. Verilog always requires the presence of an initialiser. To
support uninitialised parameters in Alogic, we would have to emit some
default value. Since we cannot enforce that the generated code, if
instantiated from Verilog will always have its parameter specified,
so it is probably better not to support uninitialized parameters.