Add special case for single state FSMs
The 'state' variable is not emitted if the FSM has only 1 sate, and all
goto statements (the one at the end of the state that goes back to
itself in particular) are treated as nops. This special case is useful
to remove some verbosity from the output, but most importantly clog2(1)
is 0, which is used to create then malformed constants and variables.