Replace BinaryArrayLookup with Slice expression
ref[a:b] is now considered an ordinary expression, though the parser
still only accepts it if 'ref' is a variable reference (which is now
either a dotted name, or a dotted name with an arbitrary number of index
expressions after it).

Also changed the parser so [a:b] slices are not accepted in array
declarations.