Emit call_stack and call_depth automatically
MakeStates can now figure out if it needs to emit a call_stack and
call_depth variable to manage function calls. If the fsm is recursive,
then the user needs to define a 'const u32 CALL_STACK_SIZE' constant,
which will be used as the call stack size. If the fsm is not recursive,
then MakeStates can derive the required stack size based on the static
call graph.