Add rewrites to support pipelines
This patch allows multiple fsms to be defined in a pipeline block.

We replace read() and write() calls with p_in and p_out calls.
We track the use of variables defined at network level as pipeline and
save/restore the appropriate ones.
We automatically connect ports from internal fsms to the network layer.
(Currently multi-connections not supported)