Fix liveness analysis of index/slice arguments properly.
The previous fix of replacing collect with collectAll went a bit too
far as the recursion on the subject of the index caused all bits to be
marked used. eg 'a[0]' would be marked with all bits of a being used,
rather than just bit 0. Fix this by selectively recursing only where
required.