Refactor so Typer doesn't rewrite any nodes
The typer should only check the tree is type sound but otherwise not
modify it. This property is useful and used by some passes when invoking
the typer in isolation to check trees are sound.

Also removed a bunch of hacks, including untyped expression folding, that
was needed by the previous implementation of the parameter
specialization pass. This makes a lot of the internals cleaner.