Don't skip nodes in the Typer.
Due to structural sharing (of parameter values in particular), it is
possible for subtrees to be already typed when the Typer encounters a
tree in the final type checking pass. If we skip these the unary '
context stack can get out of sync, so we walk these already typed
nodes but don't type check them.