Implement arbitrary entity content generation with 'gen'
Generating parametrized entities requires applying the
SpecializeParams and Generate transforms at the same time. This is now
done together in the Specialize pass, and allows correctly generating
arbitrary entity contents except for 'const' and 'param' declarations
which are not allowed to appear inside a 'gen' outside but outside a
nested entity. Recursive entities also work.