Major refactor of internal representations + improved elaboration
The internal representations of all nodes that introduce a new name are
now split into 3 flavours. A Desc (description) node is used as the source
level representation, which contains all AST elements of source level
content. The Elaborate pass converts all Desc nodes into pairs of
Decl (declaration) and Defn (definition) nodes. All symbols have either
one associated Desc node (prior to elaboration) or a pair of Decl/Defn
nodes (after elaboration). The Decl node of a symbol contains all
information required to determine the type of the symbol, and nothing
more, while the Defn node of the symbol contains all additional
information. This separation of Decl/Defn parts enforces a clear
separation throughout the compiler between transformations that change
the type of a symbol and transformations that do not. Note that the type
of a symbol now cannot really be changed. Instead, a new symbol with the
new type must be created and the old symbol substituted with the new
symbol. To facilitate this, the standard TreeTransformer has been
augmented with a new 'replace' method, which is called the first time a
symbol is referenced, and can be used to replace that symbol with a new
symbol. All further references to the old symbol will then be replaced
by identical references to the new symbol, prior to the standard tree
traversal continuing in the usual way. Note that the 'replace'
mechanism will cause the Decl of the symbol being replaced to be walked
out of order. The new Decl representation also does away with types
inside Trees (Except for the explicit ExprType to refer to a fundamental
type, and ExprCast), which means there is now no need for
TypeTransformer/TreeInTypeTransformer and related as their uses are
naturally coverted by TreeTransformer with the 'replace' mechanism.

With this refactoring, the Elaborate pass can cope with arbitrarily
complex parametrization dependencies among symbols. The only case it
fails is if there is a circular typing dependency, which is reported to
the user as an error. We can also cope with arbitrarily inter-dependent
'gen' constructs, so long as there are no circular dependencies. As a
side effect of this, the compiler can now also cope with arbitrarily
ordered definitions in any place, though some ordering is still enforced
where necessary for semantics (e.g.: assign before definition in
statements)

This patch also removes the distinction between TermSymbol and
TypeSymbol, which are now both represented as just a Symbol, and uses
the TypeType type instead to denote what used to be TypeSymbols, i.e.:
what used to be TypeSymbols are now Symbols with type TypeType(_), and
what used to be TermSymbols are now Symbols with a type other than
TypeType(_).

As there is no more need for TypeTransformer, TreeLike and
TreeLikeTransformer have been removed. The functionality of
TreeLikeTransformer has been incororated into TreeTransformer directly.
The 'enter' method of the TreeTransofrmer no longer returns Unit, but an
Option[Tree], which can be used to preempt traversal of the subtree
below the entered node by providing the replacement of the enter node
immediately.

The internal representation of 'struct' has benn completely re-worked and
is now closer to the representation of an Entity using the
DeclRecord/DefnRecord tree nodes. This means that these definitions are
now of the same class as entities, rather than being canonically
represented via a Type.