Don't use .toSource in compiler messages
TreePrintOps.toSource is not quite for generating messages anymore as it
prints symbols with an unique ID suffix, e.g.: foo@12 is the symbol
number 12 with name 'foo', as the ID is always unique but the name might
not be. We rely on the location context to highlight the message
location instead.