Lower variables into clocked processes
LoverVariables now adds up to two clocked processes (one for reset flops
and one for non-reset flops) to the entity which are used to implement
the flops, instead of them being magic in the code generation.
DefaultAssignments subsequently can remove some of these if the stored
values are not actually used.