Keep assertion logic in the assertion clocked process.
Assertions now only extract the values of modifiable symbols in their
condition, and otherwise move the whole condition expression into the
assertion block. This means only symbol to symbol direct assignments are
added in place of the assertion which require no logic. This means that
even if the Alogic compiler emits assertions, if they are subsequently
stripped, the code should only contain extra assignment statements which
are easily provable to be redundant. It also means that foreign function
calls are now deferred to assertion checking time, rather than being
in-lined with the combinational process.