Severely reduce the proliferation of the CompilerContext
There were two main reasons why cc needed to go to a lot of places:
1. Needed for sequence numbers when allocating Symbol instances in
builtins. The sequence numbers in Symbol are only used as a debugging
aid and has no other function, so builtins symbols now use independent
sequence numbers.
2. Needing access to the message buffer to report rare Error messages.
These now throw Fatal instead. The rest of these should eventually be
converted to assertions with the type checker checking the conditions of
the errors in user code.